[Update to 6.2.2]: Fix for CVE-2019-19911 CVE-2020-5310 CVE-2020-5311 CVE-2020-5312 CVE-2020-5313
Notes from https://pillow.readthedocs.io/en/stable/releasenotes/6.2.2.html :

Security
This release addresses several security problems {CVEs TBD), as well as addressing CVE-2019-19911.

CVE-2019-19911 is regarding FPX images. If an image reports that it has a large number of bands, a large amount of
resources will be used when trying to process the image. This is fixed by limiting the number of bands to those
usable by Pillow.

Buffer overruns were found when processing an SGI, PCX or FLI image. Checks have been added to prevent this.

Overflow checks have been added when calculating the size of a memory block to be reallocated in the processing of a
TIFF image.

--

Andrew Murray (16):
      Changed docker tag to 6.2.x
      Removed EOL Fedora 29
      Removed EOL Fedora 29
      Added Fedora 31
      Updated CI targets [ci skip]
      Added CentOS 8
      Updated copyright year
      Added security notes [ci skip]
      Raise an error for an invalid number of bands in FPX image
      Catch FLI buffer overrun
      Catch PCX P mode buffer overrun
      Catch SGI buffer overruns
      Overflow checks for realloc for tiff decoding
      Added release notes [ci skip]
      6.2.2 version bump
      Release notes for 6.2.2

Hugo (3):
      Format with Black 19.10b0
      Test on Python 3.8
      Test on Python 3.8

--
CVEs fixed in this build:
CVE-2019-19911
CVE-2020-5310
CVE-2020-5311
CVE-2020-5312
CVE-2020-5313